+++ performance_test/include/performance_test/execution_tasks/publisher_task.hpp
@@ -36,6 +36,8 @@ public:
   PublisherTask & operator=(const PublisherTask &) = delete;
   PublisherTask(const PublisherTask &) = delete;
 
+  void prepare();
+
   void run();
 
 private: