+++ performance_test/include/performance_test/plugin/publisher.hpp
@@ -28,6 +28,13 @@ class Publisher
 public:
   virtual ~Publisher() = default;
 
+  /// @brief Prepare for communication
+  /// This is called once, after all Publishers and Subscribers
+  /// are created, and before any messages are sent.
+  /// This is a good place to put blocking operations that do not belong in
+  /// the constructor, such as participant discovery.
+  virtual void prepare() {}
+
   virtual void publish_copy(
     const TimestampProvider & timestamp_provider,
     std::uint64_t sample_id) = 0;